matchers: Split out WithinGeometryMatcher into its own header
