+++ matchers/meson.build
@@ -27,3 +27,23 @@ mathematical_model_matcher_dep = declare_dependency(
   include_directories: [ include_directories('.') ],
 )
 
+within_geometry_matcher_sources = [
+  'within_geometry_matcher.h',
+  'within_geometry_matcher.cpp'
+]
+
+within_geometry_matcher_lib = static_library(
+  'within_geometry_matcher',
+  within_geometry_matcher_sources,
+  dependencies: [
+    gtest_project.get_variable('gtest_dep'),
+    gtest_project.get_variable('gmock_dep'),
+    wobbly_dep
+  ],
+  install: false
+)
+
+within_geometry_matcher_dep = declare_dependency(
+  link_with: within_geometry_matcher_lib,
+  include_directories: [ include_directories('.') ],
+)
