+++ matchers/within_geometry_matcher.cpp
@@ -1,0 +1,10 @@
+/*
+ * matchers/within_geometry_matcher.cpp
+ *
+ * Provides utilities to match types satisfying the
+ * Dimension trait to ensure they are within
+ * a PointBox.
+ *
+ * See LICENCE.md for Copyright information
+ */
+#include "within_geometry_matcher.h"
